backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
pytest
pytest-asyncio 